Updates to build this demo with recent changes
This tree was a little bit dated, so this is an update that will make
sure this demo builds again on latest today (2015-12-15).

There is a patch:
  0001-Patch-for-file-inclusion-of-Comcast-demo-TA.patch
That should be applied on optee/build.git (see repo instructions at
optee_os/README.md at GitHub).

The is also a file(s) for qemu and fvp:
  filelist-comcast_demo_{qemu,fvp}_fs.txt
Note that this file contains hardcoded absolute paths that must be
updated according to your own computer/paths.

To build this, you specify the target (either qemu or fvp), for example:
  $ TARGET=qemu ./build_comcast_cryptoapi_ta_example

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>